requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15,<3.17
celery>=5.3,<6.0
redis>=5.0,<6.0
mssql-django>=1.5,<2.0
pyodbc>=5.1,<6.0
python-dotenv>=1.0,<2.0
